pyproject.toml
[project]
name = "collector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-flask>=1.3.0",
    "pytest-asyncio>=0.23.6",
    "factory-boy>=3.3.0",
    "testcontainers[postgres]>=4.4.0",
    "playwright>=1.43.0",
]


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "E", "F", "Q000"]
ignore = []

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
